+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.6.1.dev
 
 * Fix: MOVE failing with URL-encoded destination header
+* Improve: add workaround to remove empty lines in item to avoid reject by vobject parser
 
 ## 3.6.0